requirements.txt
pyinstaller
docopt
